comps/vector-store/app/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
durationpy==0.10
fastapi==0.120.3
google-auth==2.42.1
h11==0.16.0
idna==3.11
kubernetes==34.1.0
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.3
pydantic-core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.3.0
uvicorn==0.38.0
websocket-client==1.9.0
